Add inital dashboard page
