+++ app/controllers/public_controller.rb
@@ -1,0 +1,6 @@
+class PublicController < ApplicationController
+  before_action :logged_in_user
+  def index
+
+  end
+end
