+++ app/controllers/users_controller.rb
@@ -53,14 +53,6 @@ class UsersController < ApplicationController
     params.require(:user).permit(:name, :email, :password, :password_confirmation)
   end
 
-  def logged_in_user
-    unless logged_in?
-      store_location
-      flash[:danger] = "Please log in."
-      redirect_to login_url, status: :see_other
-    end
-  end
-
   def correct_user
     @user = User.find(params[:id])
     redirect_to(root_url, status: :see_other) unless current_user?(@user)
