+++ app/helpers/application_helper.rb
@@ -1,2 +1,7 @@
 module ApplicationHelper
+  def show_svg(path)
+    File.open("app/assets/images/#{path}", "rb") do |file|
+      raw file.read
+    end
+  end
 end
