+++ app/helpers/public_helper.rb
@@ -1,0 +1,2 @@
+module PublicHelper
+end
