+++ app/helpers/users_helper.rb
@@ -1,2 +1,9 @@
 module UsersHelper
+  def logged_in_user
+    unless logged_in?
+      store_location
+      flash[:danger] = "Please log in."
+      redirect_to login_url, status: :see_other
+    end
+  end
 end
