+++ app/views/layouts/_header.html.erb
@@ -1,0 +1,30 @@
+<div class="navbar bg-base-100">
+  <div class="flex-1">
+    <%= link_to "PocketWealth", home_path, class: "btn btn-ghost text-xl" %>
+  </div>
+  <div class="flex-none gap-2">
+      <% unless current_user.nil? %>
+      <div class="dropdown dropdown-end">
+        <div tabindex="0" role="button" class="btn btn-ghost btn-circle avatar">
+          <div class="w-10 rounded-full">
+            <img
+              alt="Tailwind CSS Navbar component"
+              src="https://img.daisyui.com/images/stock/photo-1534528741775-53994a69daeb.webp" />
+          </div>
+        </div>
+        <ul
+          tabindex="0"
+          class="menu menu-sm dropdown-content bg-base-100 rounded-box z-[1] mt-3 w-52 p-2 shadow">
+          <li>
+            <a class="justify-between">
+              <%= link_to "Profile", user_path(current_user) %>
+            </a>
+          </li>
+          <li><a>Settings</a></li>
+          <li><%= link_to "Log out", logout_path,
+                          data: { "turbo-method": :delete } %></li>
+        </ul>
+      <% end %>
+    </div>
+  </div>
+</div>
