+++ app/views/public/index.html.erb
@@ -1,0 +1,18 @@
+<div class="flex flex-col">
+  <div class="drawer drawer-open">
+    <input id="my-drawer-2" type="checkbox" class="drawer-toggle" />
+    <div class="drawer-side">
+      <label for="my-drawer-2" aria-label="close sidebar" class="drawer-overlay"></label>
+      <ul class="menu bg-base-200 text-base-content min-h-full w-80 p-4">
+        <!-- Sidebar content here -->
+        <li><a class="p-3 bg-primary mb-5 font-bold">Accounts</a></li>
+        <li><a class="p-3 bg-primary mb-5 font-bold">Transactions</a></li>
+        <li><a class="p-3 bg-primary mb-5 font-bold">Stocks</a></li>
+        <li><a class="p-3 bg-primary mb-5 font-bold">Other</a></li>
+      </ul>
+    </div>
+  </div>
+  <div>
+
+  </div>
+</div>
