+++ config/application.rb
@@ -23,5 +23,6 @@ module PocketWealth
     #
     # config.time_zone = "Central Time (US & Canada)"
     # config.eager_load_paths << Rails.root.join("extras")
+    config.assets.css_compressor = nil
   end
 end
