+++ config/routes.rb
@@ -1,11 +1,12 @@
 Rails.application.routes.draw do
-  root "users#index"
+  root "public#index"
   get "sessions/new"
   get  "/signup",  to: "users#new"
   get  "/login",  to: "sessions#new"
   post "/login",  to: "sessions#create"
   delete "/logout", to: "sessions#destroy"
   resources :users
+  get "/home",  to: "public#index"
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
