+++ package.json
@@ -1,5 +1,6 @@
 {
   "devDependencies": {
-    "daisyui": "^4.12.10"
+    "daisyui": "^4.12.10",
+    "tailwindcss": "^3.4.7"
   }
 }
